refactor: Auth/Onboarding 디렉토리명 PascalCase 적용 #11
