+++ src/screens/Auth/NicknameSetupScreen.tsx
